Optimised code and design for different screen sizes
Deleted Temporary Lock View controller as it was not used anywhere in
the project.

Created a single method to delete selected images and use the function
when user click on Trash Button or Unlock button.

Changed the Alert messages that appears in delete alert.

Add constraints into Main.storyboard for Authentication, Photo Album
and PhotoSlider view so it looks better after orientation. It also let
the components in the app to arrange themselves automatically according
to the screen size.

Things that still need to be done:
- Arrange Cells in UICollectionView properly
- Show Navigation bar after tap on the Photo in PhotoSlider
- Customise Page controls in PhotoSlider