Multiple Photos selection works better
I removed “self” keyword from the places where it was not required.

I have put one condition to check if cell is initialised or not before
deselecting it in collection view.

Deleted the override method which was not executing anytime to deselect
cell at indexpath selected by user.

Made changes in cellForItemAtIndexPath method which makes decision
about the border color of cell, because UICollectionView reloads its
data while user scroll it, so now it shows orange border to the cell
which were selected by user. Cell’s image Rasterize property is set
which help in smooth scrolling of UICollectionView.

didSelectItemAtIndexPath method is responsible for selecting and
deselecting the cell selected by user.

Things that needs to be done:
In PhotoSlider one bar comes and hide itself after first tap on photo.
I don’t know if I had ever created or used it.