Add a condition before getting shared folder url (ShareExtension)
I found that iVault sometimes crashes after several time of interval.
The problem was that my provisioning profile was getting expired in
that time and App was not able to get ShareExtension’s container URL
and get crashed because of nil value in
copySharedImagesToDocumentDirectory().

As a solution I just put one condition before whole code so it don’t
crashes if it could’t get path of shared container, it won’t load
images shared through other apps in the photo album though.

But it is better to do this for good User experience.